(PC-13649)[API] refactor: move functions in dms_repository
